third_party/WebKit/LayoutTests/fast/dom/MutationObserver/script-append-expected.txt

Test that an appended script does NOT invoke MutationObserver callbacks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutationsDelivered is false
PASS scriptDidRun is false
PASS scriptDidRun is true
PASS mutationsDelivered is false
PASS successfullyParsed is true

TEST COMPLETE
